Change the argument to be a variant
In preparation to allow more than one argument type, use a `variant`
which holds only a 64-bit integer for now. This means there should be no
change in behavior, but we are prepared to add more possible argument
types.

In TestUtilities.cpp we added a include header so that output of
`variant` with fmtlib would work.